translator-bot/pyproject.toml
[project]
name = "bilingual-bot"
version = "0.1.0"
description = "Telegram bot that converts Ayurvedic exam quiz DOCX/XLSX files into bilingual English+Hindi versions using Claude API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=20.7",
    "anthropic>=0.25.0",
    "openai>=1.0.0",
    "lxml>=5.1.0",
    "openpyxl>=3.1.2",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bot"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
